+++ commands/vm-autoscaling/action/action.go
@@ -39,10 +39,12 @@ func Root() *core.Command {
 var (
 	allJSONPaths = map[string]string{
 		"ActionId": "id",
+		"GroupId":  "href",
 	}
 
 	allCols = []string{
 		"ActionId",
+		"GroupId",
 	}
 
 	defaultCols = allCols